requirements.txt
aiogram~=3.18.0
requests
httpx~=0.28.1
psycopg2-binary~=2.9.9
loguru~=0.7.3

python-dateutil~=2.9.0.post0
pip
attrs~=25.1.0
wheel~=0.37.0
six~=1.15.0
